+++ fluorender/FluoRender/FLIVR/VertexArray.h
@@ -411,10 +411,10 @@ namespace FLIVR
 	{
 		if (border)
 			glDrawElements(GL_LINE_LOOP, 4, GL_UNSIGNED_INT,
-				reinterpret_cast<const GLvoid*>(__int64(plane)));
+				reinterpret_cast<const GLvoid*>((long long)(plane)));
 		else
 			glDrawElements(GL_TRIANGLE_STRIP, 4, GL_UNSIGNED_INT,
-				reinterpret_cast<const GLvoid*>(__int64(plane)));
+				reinterpret_cast<const GLvoid*>((long long)(plane)));
 	}
 
 	inline void VertexArray::draw_grid()
